fixed mob AIs all following the same walk pattern
