+++ core/src/miniventure/game/world/entity/mob/MovementPattern.java
@@ -2,9 +2,11 @@ package miniventure.game.world.entity.mob;
 
 import com.badlogic.gdx.math.Vector2;
 
-@FunctionalInterface
+//@FunctionalInterface
 public interface MovementPattern {
 	
 	Vector2 move(float delta, MobAi mob);
+	
+	MovementPattern copy();
 	
 }
